% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smpk.R
\name{smpk}
\alias{smpk}
\title{Fit SMPK model}
\usage{
smpk(Y, C, prior = NULL, pmc = NULL, state = NULL)
}
\arguments{
\item{Y}{Matrix of the data. Each row represents an observation.}

\item{C}{Vector of the group label of each observation. Labels are integers starting from 1.}

\item{prior}{A list giving the prior information. If unspecified, a default prior is used.
The list includes the following hyparameters:
\code{K} Number of mixture components.
\code{epsilon_range} Vector with minimum and maximum values for \code{epsilon}.
\code{merge_step} Introduce step to merge mixture components with small KL divergence. Default is \code{merge_step = TRUE}.
\code{merge_par} Parameter controlling merging radius. Default is \code{merge_par = 0.1}.}
}
\description{
This function generates a sample from the posterior of CREMID.
}
